Migrate `:data:model` and `:data:remote` modules to Kotlin Multiplatform
