+++ data/models/build.gradle.kts
@@ -2,5 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 plugins {
-  id("red.razvan.restcountries.kotlin.jvm")
+  id("red.razvan.restcountries.kotlin.multiplatform")
+}
+
+kotlin {
+  jvm()
+  iosArm64()
+  iosSimulatorArm64()
 }
