+++ data/models/src/commonMain/kotlin/red/razvan/restcountries/data/models/CountryId.kt
@@ -1,0 +1,9 @@
+// Copyright 2025 Răzvan Roșu
+// SPDX-License-Identifier: Apache-2.0
+
+package red.razvan.restcountries.data.models
+
+import kotlin.jvm.JvmInline
+
+@JvmInline
+value class CountryId(val value: String)
