+++ data/models/src/commonMain/kotlin/red/razvan/restcountries/data/models/CountryListItem.kt
