+++ data/models/src/commonMain/kotlin/red/razvan/restcountries/data/models/DetailedCountry.kt
