--- data/models/src/main/kotlin/red/razvan/restcountries/data/models/CountryId.kt
@@ -1,7 +1,0 @@
-// Copyright 2025 Răzvan Roșu
-// SPDX-License-Identifier: Apache-2.0
-
-package red.razvan.restcountries.data.models
-
-@JvmInline
-value class CountryId(val value: String)
