+++ data/remote/build.gradle.kts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 plugins {
-  id("red.razvan.restcountries.kotlin.android")
+  id("red.razvan.restcountries.kotlin.multiplatform")
   id("red.razvan.restcountries.android.library")
   alias(libs.plugins.kotlin.serialization)
 }
@@ -11,19 +11,46 @@ android {
   namespace = "red.razvan.restcountries.data.remote"
 }
 
-dependencies {
-  implementation(libs.kotlinx.coroutines.core)
-  implementation(libs.ktor.serialization.kotlinx.json)
-  implementation(libs.ktor.client.core)
-  implementation(libs.ktor.client.okhttp)
-  implementation(libs.ktor.client.logging)
-  implementation(libs.ktor.client.contentnegotiation)
-  api(libs.koin.android)
+kotlin {
+  androidTarget()
+  iosArm64()
+  iosSimulatorArm64()
 
-  testImplementation(libs.ktor.client.mock)
-  testImplementation(libs.assertk)
-  testImplementation(libs.assertk.coroutines)
-  testImplementation(libs.kotlinx.coroutines.test)
-  testImplementation(libs.koin.test)
-  testImplementation(libs.koin.test.junit4)
+  sourceSets {
+    val commonMain by getting {
+      dependencies {
+        implementation(libs.kotlinx.coroutines.core)
+        implementation(libs.ktor.serialization.kotlinx.json)
+        implementation(libs.ktor.client.core)
+        implementation(libs.ktor.client.logging)
+        implementation(libs.ktor.client.contentnegotiation)
+        implementation(projects.commons.intellijAnnotations)
+        api(libs.koin.core)
+      }
+    }
+
+    val iosMain by getting {
+      dependencies {
+        implementation(libs.ktor.client.darwin)
+      }
+    }
+
+    val commonTest by getting {
+      dependencies {
+        implementation(libs.kotlin.test)
+        implementation(libs.ktor.client.mock)
+        implementation(libs.assertk)
+        implementation(libs.assertk.coroutines)
+        implementation(libs.kotlinx.coroutines.test)
+        implementation(libs.koin.test)
+      }
+    }
+
+    val androidMain by getting {
+      dependencies {
+        api(libs.koin.android)
+        implementation(libs.ktor.client.okhttp)
+      }
+    }
+  }
 }
