+++ data/remote/src/commonMain/kotlin/red/razvan/restcountries/data/remote/DetailedCountry.kt
