+++ data/remote/src/commonMain/kotlin/red/razvan/restcountries/data/remote/PlatformModule.kt
@@ -1,0 +1,8 @@
+// Copyright 2025 Răzvan Roșu
+// SPDX-License-Identifier: Apache-2.0
+
+package red.razvan.restcountries.data.remote
+
+import org.koin.core.module.Module
+
+internal expect val PlatformModule: Module
